Move UserConfig to its own cpp/h
