+++ source/ConfigFiles.h
@@ -15,180 +15,9 @@ static bool operator==(Array<T> a1, Array<T> a2) {
 	return !(a1 != a2);
 }
 
-/**Class for managing user configuration*/
-class UserConfig {
-public:
-    String			id					= "anon";						///< Subject ID (as recorded in output DB)
-    double			mouseDPI			= 800.0;						///< Mouse DPI setting
-	double			mouseDegPerMm		= 2.824;						///< Mouse sensitivity, reported as degree per mm
-	Vector2			turnScale			= Vector2(1.0f, 1.0f);			///< Turn scale for player, can be used to invert controls in either direction
-	bool			invertY				= false;						///< Extra flag for Y-invert (duplicates turn scale, but very common)
-	Vector2			scopeTurnScale		= Vector2(0.0f, 0.0f);			///< Scoped turn scale (0's imply default scaling)
-
-	int				currentSession		= 0;							///< Currently selected session
-		
-	int				reticleIndex		= 39;							///< Reticle to show for this user
-	Array<float>	reticleScale		= { 1.0f, 1.0f };				///< Scale for the user's reticle
-	Array<Color4>	reticleColor		= {Color4(1.0, 0.0, 0.0, 1.0),	///< Color for the user's reticle
-										Color4(1.0, 0.0, 0.0, 1.0)};	
-	float			reticleChangeTimeS	= 0.3f;							///< Time for reticle to contract after expand on shot (in seconds)
 
 
-	UserConfig() {};
 
-	/** Load from Any */
-    UserConfig(const Any& any) {
-		// for loading old user configs
-		double cmp360 = 12.75;
-		bool foundMouseDegPerMm = false;
-
-        int settingsVersion = 1; // used to allow different version numbers to be loaded differently
-        AnyTableReader reader(any);
-		reader.getIfPresent("settingsVersion", settingsVersion);
-        switch (settingsVersion) {
-        case 1:
-            reader.getIfPresent("id", id);
-            reader.getIfPresent("mouseDPI", mouseDPI);
-            foundMouseDegPerMm = reader.getIfPresent("mouseDegPerMillimeter", mouseDegPerMm);
-			reader.getIfPresent("cmp360", cmp360);
-			reader.getIfPresent("reticleIndex", reticleIndex);
-			reader.getIfPresent("reticleScale", reticleScale);
-			reader.getIfPresent("reticleColor", reticleColor);
-			reader.getIfPresent("reticleChangeTime", reticleChangeTimeS);
-			reader.getIfPresent("turnScale", turnScale);
-			reader.getIfPresent("invertY", invertY);
-			reader.getIfPresent("scopeTurnScale", scopeTurnScale);
-			break;
-        default:
-            debugPrintf("Settings version '%d' not recognized in UserConfig.\n", settingsVersion);
-            break;
-        }
-
-		// Set mouseDPmm if not found
-		if (!foundMouseDegPerMm) {
-			mouseDegPerMm = 36.0 / cmp360;
-		}
-    }
-	
-	/** Serialize to Any */
-	Any toAny(const bool forceAll=true) const {
-		UserConfig def;
-		Any a(Any::TABLE);
-		a["id"] = id;										// Include subject ID
-		a["mouseDPI"] = mouseDPI;							// Include mouse DPI
-		a["mouseDegPerMillimeter"] = mouseDegPerMm;						// Include sensitivity
-		if (forceAll || def.reticleIndex != reticleIndex)				a["reticleIndex"] = reticleIndex;
-		if (forceAll || def.reticleScale != reticleScale)				a["reticleScale"] = reticleScale;
-		if (forceAll || def.reticleColor != reticleColor)				a["reticleColor"] = reticleColor;
-		if (forceAll || def.reticleChangeTimeS != reticleChangeTimeS)	a["reticleChangeTime"] = reticleChangeTimeS;
-		if (forceAll || def.turnScale != turnScale)						a["turnScale"] = turnScale;
-		if (forceAll || def.invertY != invertY)							a["invertY"] = invertY;
-		if (forceAll || def.scopeTurnScale != scopeTurnScale)			a["scopeTurnScale"] = scopeTurnScale;
-		return a;
-	}
-
-	// Define not equal operator for comparison
-	bool operator==(const UserConfig& other) const {
-		bool eq = id == other.id && mouseDegPerMm == other.mouseDegPerMm && reticleIndex == other.reticleIndex &&
-			reticleScale == other.reticleScale && reticleColor == other.reticleColor && reticleChangeTimeS == other.reticleChangeTimeS && 
-			turnScale == other.turnScale && invertY == other.invertY && scopeTurnScale == other.scopeTurnScale;
-		return eq;
-	}
-};
-
-/** Class for loading a user table and getting user info */
-class UserTable {
-public:
-
-	bool					requireUnique = true;			///< Require users to be unique by ID
-	UserConfig				defaultUser;					///< Default user settings to use for new user
-	Array<UserConfig>		users = {};						///< A list of valid users
-
-	UserTable() {};
-
-	/** Load from Any */
-	UserTable(const Any& any) {
-		int settingsVersion = 1;
-		AnyTableReader reader(any);
-		reader.getIfPresent("settingsVersion", settingsVersion);
-
-		switch (settingsVersion) {
-		case 1:
-			reader.getIfPresent("requireUnique", requireUnique);
-			reader.getIfPresent("defaultUser", defaultUser);
-			reader.get("users", users, "Issue in the (required) \"users\" array in the user config file!");
-			// Unique user check (if required)
-			if (requireUnique) {
-				const Array<String> userIds = getIds();
-				for (String id : userIds) {
-					if (userIds.findIndex(id) != userIds.rfindIndex(id)) {
-						throw "Multiple users with the same ID (\"" + id + "\") specified in the user config file!";
-					}
-				}
-			}
-			if (users.size() == 0) {
-				throw "At least 1 user must be specified in the \"users\" array within the user configuration file!";
-			}
-			break;
-		default:
-			debugPrintf("Settings version '%d' not recognized in UserTable.\n", settingsVersion);
-			break;
-		}
-	}
-
-	/** Serialize to Any */
-	Any toAny(const bool forceAll = true) const {
-		Any a(Any::TABLE);
-		a["settingsVersion"] = 1;						///< Create a version 1 file
-		a["users"] = users;								///< Include updated subject table
-		return a;
-	}
-
-	/** Get a user config based on a user ID */
-	shared_ptr<UserConfig> getUserById(String id) {
-		for (UserConfig user : users) {
-			if (!user.id.compare(id)) return std::make_shared<UserConfig>(user);
-		}
-		return nullptr;
-	}
-
-	/** Simple rotine to get the UserTable Any structure from file */
-	static UserTable load(const String& filename) {
-		// Create default UserConfig file
-		if (!FileSystem::exists(System::findDataFile(filename, false))) { // if file not found, generate a default user config table
-			UserTable defTable = UserTable();
-			defTable.users.append(UserConfig());			// Append one default user
-			defTable.save(filename);						// Save the .any file
-			return defTable;
-		}
-		return Any::fromFile(System::findDataFile(filename));
-	}
-
-	inline void save(const String& filename) { toAny().save(filename); }
-
-	/** Get an array of user IDs */
-	Array<String> getIds() {
-		Array<String> ids;
-		for (UserConfig user : users) ids.append(user.id);
-		return ids;
-	}
-
-	/** Get the index of the current user from the user table */
-	int getUserIndex(String userId) {
-		for (int i = 0; i < users.length(); ++i) {
-			if (!users[i].id.compare(userId)) return i;
-		}
-		// return the first user by default
-		return 0;
-	}
-
-	/** Print the user table to the log */
-	void printToLog() {
-		for (UserConfig user : users) {
-			logPrintf("\tUser ID: %s, sensitivity = %f deg/mm, mouseDPI = %d\n", user.id.c_str(), user.mouseDegPerMm, user.mouseDPI);
-		}
-	}
-};
 
 /** Class for handling user status */
 class UserSessionStatus {
