+++ source/GuiElements.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <G3D/G3D.h>
 #include "ConfigFiles.h"
+#include "UserConfig.h"
 #include "TargetEntity.h"
 
 class FPSciApp;
