+++ vs/FPSci.lib.vcxproj.filters
@@ -72,6 +72,9 @@
     <ClInclude Include="..\source\SystemConfig.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\source\UserConfig.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\source\sqlHelpers.cpp">
@@ -114,6 +117,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\source\SystemConfig.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\source\UserConfig.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
